Cast audio to float32 in DistillNeuCodec forward to avoid bfloat16/numpy incompatibility

+++ distill_neucodec/pytorch/loader.py
@@ -93,7 +93,8 @@ class ModelLoader(ForgeModel):
                 self.model = model
 
             def forward(self, audio):
-                return self.model.encode_code(audio)
+                # Feature extractor cannot handle bfloat16; convert to float32
+                return self.model.encode_code(audio.float())
 
         model = DistillNeuCodecWrapper(codec)
         model.eval()
